% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dt_dates.R
\docType{data}
\name{dt_dates}
\alias{dt_dates}
\title{Anonymized sample data}
\format{a data table with dates}
\usage{
data(dt_dates)
}
\description{
Anonymized sample data
}
\examples{
data(dt_dates)
}
\author{
Jiena Gu McLellan, 2020-05-26
}
\keyword{datasets}
