% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dt_groups.R
\docType{data}
\name{dt_groups}
\alias{dt_groups}
\title{Anonymized sample data}
\format{a data table with groups}
\usage{
data(dt_groups)
}
\description{
Anonymized sample data
}
\examples{
data(dt_groups)
}
\author{
Jiena Gu McLellan, 2020-05-26
}
\keyword{datasets}
