% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dt_long.R
\docType{data}
\name{dt_long}
\alias{dt_long}
\title{Anonymized sample data}
\format{a data table in long format}
\usage{
data(dt_long)
}
\description{
Anonymized sample data
}
\examples{
data(dt_long)
}
\author{
Jiena Gu McLellan, 2020-05-26
}
\keyword{datasets}
