% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dt_missing.R
\docType{data}
\name{dt_missing}
\alias{dt_missing}
\title{Anonymized sample data}
\format{a data table with missing values}
\usage{
data(dt_missing)
}
\description{
Anonymized sample data
}
\examples{
data(dt_missing)
}
\author{
Jiena Gu McLellan, 2020-05-26
}
\keyword{datasets}
