% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dt_values.R
\docType{data}
\name{dt_values}
\alias{dt_values}
\title{Anonymized sample data}
\format{a data table with values}
\usage{
data(dt_values)
}
\description{
Anonymized sample data
}
\examples{
data(dt_values)
}
\author{
Jiena Gu McLellan, 2020-05-26
}
\keyword{datasets}
