% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_str_reshape.R
\name{fill_NA_with}
\alias{fill_NA_with}
\title{Fill missing values}
\usage{
fill_NA_with(dt, fill_cols, fill_value)
}
\arguments{
\item{dt}{input data table}

\item{fill_cols}{filter by this columns}

\item{fill_value}{fill NA with this value}
}
\value{
data table which NAs are filled
}
\description{
Fill missing values
}
\examples{
data("dt_missing")
fill_NA_with(dt_missing, fill_cols = c("Full_name"), fill_value = "pending")
}
