% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_str_reshape.R
\name{reshape_longer}
\alias{reshape_longer}
\title{Title}
\usage{
reshape_longer(
  dt,
  keep_cols,
  by_pattern,
  label_cols,
  value_cols,
  fill_NA_with = NULL
)
}
\arguments{
\item{fill_NA_with}{}
}
\value{

}
\description{
Title
}
